moved detector and eta-omega out of instrument file
