+++ hexrd/instrument/instrument.py
@@ -43,9 +43,7 @@ import numpy as np
 from scipy import ndimage
 from scipy.linalg.matfuncs import logm
 
-from hexrd.gridutil import cellIndices, make_tolerance_grid
-from hexrd import matrixutil as mutil
-from hexrd.valunits import valWUnit
+from hexrd.gridutil import make_tolerance_grid
 from hexrd.xrd.transforms_CAPI import anglesToGVec, \
                                       detectorXYToGvec, \
                                       gvecToDetectorXY, \
@@ -57,13 +55,7 @@ from hexrd.xrd.transforms_CAPI import anglesToGVec, \
                                       rowNorm, \
                                       validateAngleRanges
 from hexrd.xrd import xrdutil
-from hexrd.xrd.crystallography import PlaneData
 from hexrd import constants as ct
-
-# from hexrd.utils.progressbar import ProgressBar, Bar, ETA, ReverseBar
-
-# FIXME: distortion kludge
-from hexrd.xrd.distortion import GE_41RT  # BAD, VERY BAD!!!
 
 from skimage.draw import polygon
 
@@ -983,1278 +975,3 @@ class HEDMInstrument(object):
     """def fit_grain(self, grain_params, data_dir='results'):"""
 
     pass  # end class: HEDMInstrument
-
-
-class PlanarDetector(object):
-    """
-    base class for 2D planar, rectangular row-column detector
-    """
-
-    __pixelPitchUnit = 'mm'
-
-    def __init__(self,
-                 rows=2048, cols=2048,
-                 pixel_size=(0.2, 0.2),
-                 tvec=np.r_[0., 0., -1000.],
-                 tilt=ct.zeros_3,
-                 name='default',
-                 bvec=ct.beam_vec,
-                 evec=ct.eta_vec,
-                 saturation_level=None,
-                 panel_buffer=None,
-                 roi=None,
-                 distortion=None):
-        """
-        panel buffer is in pixels...
-
-        """
-        self._name = name
-
-        self._rows = rows
-        self._cols = cols
-
-        self._pixel_size_row = pixel_size[0]
-        self._pixel_size_col = pixel_size[1]
-
-        self._saturation_level = saturation_level
-
-        if panel_buffer is None:
-            self._panel_buffer = 20*np.r_[self._pixel_size_col,
-                                          self._pixel_size_row]
-
-        self._roi = roi
-
-        self._tvec = np.array(tvec).flatten()
-        self._tilt = np.array(tilt).flatten()
-
-        self._bvec = np.array(bvec).flatten()
-        self._evec = np.array(evec).flatten()
-
-        self._distortion = distortion
-
-        return
-
-    # detector ID
-    @property
-    def name(self):
-        return self._name
-
-    @name.setter
-    def name(self, s):
-        assert isinstance(s, (str, unicode)), "requires string input"
-        self._name = s
-
-    # properties for physical size of rectangular detector
-    @property
-    def rows(self):
-        return self._rows
-
-    @rows.setter
-    def rows(self, x):
-        assert isinstance(x, int)
-        self._rows = x
-
-    @property
-    def cols(self):
-        return self._cols
-
-    @cols.setter
-    def cols(self, x):
-        assert isinstance(x, int)
-        self._cols = x
-
-    @property
-    def pixel_size_row(self):
-        return self._pixel_size_row
-
-    @pixel_size_row.setter
-    def pixel_size_row(self, x):
-        self._pixel_size_row = float(x)
-
-    @property
-    def pixel_size_col(self):
-        return self._pixel_size_col
-
-    @pixel_size_col.setter
-    def pixel_size_col(self, x):
-        self._pixel_size_col = float(x)
-
-    @property
-    def pixel_area(self):
-        return self.pixel_size_row * self.pixel_size_col
-
-    @property
-    def saturation_level(self):
-        return self._saturation_level
-
-    @saturation_level.setter
-    def saturation_level(self, x):
-        if x is not None:
-            assert np.isreal(x)
-        self._saturation_level = x
-
-    @property
-    def panel_buffer(self):
-        return self._panel_buffer
-
-    @panel_buffer.setter
-    def panel_buffer(self, x):
-        """if not None, a buffer in mm (x, y)"""
-        if x is not None:
-            assert len(x) == 2 or x.ndim == 2
-        self._panel_buffer = x
-
-    @property
-    def roi(self):
-        return self._roi
-
-    @roi.setter
-    def roi(self, vertex_array):
-        """
-        vertex array must be
-
-        [[r0, c0], [r1, c1], ..., [rn, cn]]
-
-        and have len >= 3
-
-        does NOT need to repeat start vertex for closure
-        """
-        assert len(vertex_array) >= 3
-        self._roi = vertex_array
-
-    @property
-    def row_dim(self):
-        return self.rows * self.pixel_size_row
-
-    @property
-    def col_dim(self):
-        return self.cols * self.pixel_size_col
-
-    @property
-    def row_pixel_vec(self):
-        return self.pixel_size_row*(0.5*(self.rows-1)-np.arange(self.rows))
-
-    @property
-    def row_edge_vec(self):
-        return self.pixel_size_row*(0.5*self.rows-np.arange(self.rows+1))
-
-    @property
-    def col_pixel_vec(self):
-        return self.pixel_size_col*(np.arange(self.cols)-0.5*(self.cols-1))
-
-    @property
-    def col_edge_vec(self):
-        return self.pixel_size_col*(np.arange(self.cols+1)-0.5*self.cols)
-
-    @property
-    def corner_ul(self):
-        return np.r_[-0.5 * self.col_dim,  0.5 * self.row_dim]
-
-    @property
-    def corner_ll(self):
-        return np.r_[-0.5 * self.col_dim, -0.5 * self.row_dim]
-
-    @property
-    def corner_lr(self):
-        return np.r_[0.5 * self.col_dim, -0.5 * self.row_dim]
-
-    @property
-    def corner_ur(self):
-        return np.r_[0.5 * self.col_dim,  0.5 * self.row_dim]
-
-    @property
-    def tvec(self):
-        return self._tvec
-
-    @tvec.setter
-    def tvec(self, x):
-        x = np.array(x).flatten()
-        assert len(x) == 3, 'input must have length = 3'
-        self._tvec = x
-
-    @property
-    def tilt(self):
-        return self._tilt
-
-    @tilt.setter
-    def tilt(self, x):
-        assert len(x) == 3, 'input must have length = 3'
-        self._tilt = np.array(x).squeeze()
-
-    @property
-    def bvec(self):
-        return self._bvec
-
-    @bvec.setter
-    def bvec(self, x):
-        x = np.array(x).flatten()
-        assert len(x) == 3 and sum(x*x) > 1-ct.sqrt_epsf, \
-            'input must have length = 3 and have unit magnitude'
-        self._bvec = x
-
-    @property
-    def evec(self):
-        return self._evec
-
-    @evec.setter
-    def evec(self, x):
-        x = np.array(x).flatten()
-        assert len(x) == 3 and sum(x*x) > 1-ct.sqrt_epsf, \
-            'input must have length = 3 and have unit magnitude'
-        self._evec = x
-
-    @property
-    def distortion(self):
-        return self._distortion
-
-    @distortion.setter
-    def distortion(self, x):
-        """
-        Probably should make distortion a class...
-        ***FIX THIS***
-        """
-        assert len(x) == 2 and hasattr(x[0], '__call__'), \
-            'distortion must be a tuple: (<func>, params)'
-        self._distortion = x
-
-    @property
-    def rmat(self):
-        return makeDetectorRotMat(self.tilt)
-
-    @property
-    def normal(self):
-        return self.rmat[:, 2]
-
-    @property
-    def beam_position(self):
-        """
-        returns the coordinates of the beam in the cartesian detector
-        frame {Xd, Yd, Zd}.  NaNs if no intersection.
-        """
-        output = np.nan * np.ones(2)
-        b_dot_n = np.dot(self.bvec, self.normal)
-        if np.logical_and(
-            abs(b_dot_n) > ct.sqrt_epsf,
-            np.sign(b_dot_n) == -1
-        ):
-            u = np.dot(self.normal, self.tvec) / b_dot_n
-            p2_l = u*self.bvec
-            p2_d = np.dot(self.rmat.T, p2_l - self.tvec)
-            output = p2_d[:2]
-        return output
-
-    # ...memoize???
-    @property
-    def pixel_coords(self):
-        pix_i, pix_j = np.meshgrid(
-            self.row_pixel_vec, self.col_pixel_vec,
-            indexing='ij')
-        return pix_i, pix_j
-
-    # ...memoize???
-    @property
-    def pixel_angles(self):
-        pix_i, pix_j = self.pixel_coords
-        xy = np.ascontiguousarray(
-            np.vstack([
-                pix_j.flatten(), pix_i.flatten()
-                ]).T
-            )
-        angs, g_vec = detectorXYToGvec(
-            xy, self.rmat, ct.identity_3x3,
-            self.tvec, ct.zeros_3, ct.zeros_3,
-            beamVec=self.bvec, etaVec=self.evec)
-        del(g_vec)
-        tth = angs[0].reshape(self.rows, self.cols)
-        eta = angs[1].reshape(self.rows, self.cols)
-        return tth, eta
-
-    def config_dict(self, chi, t_vec_s, sat_level=None):
-        """
-        """
-        if sat_level is None:
-            sat_level = self.saturation_level
-
-        t_vec_s = np.atleast_1d(t_vec_s)
-
-        d = dict(
-            detector=dict(
-                transform=dict(
-                    tilt_angles=self.tilt.tolist(),
-                    t_vec_d=self.tvec.tolist(),
-                ),
-                pixels=dict(
-                    rows=self.rows,
-                    columns=self.cols,
-                    size=[self.pixel_size_row, self.pixel_size_col],
-                ),
-            ),
-            oscillation_stage=dict(
-                chi=chi,
-                t_vec_s=t_vec_s.tolist(),
-            ),
-        )
-
-        if sat_level is not None:
-            d['detector']['saturation_level'] = sat_level
-
-        if self.distortion is not None:
-            """...HARD CODED DISTORTION! FIX THIS!!!"""
-            dist_d = dict(
-                function_name='GE_41RT',
-                parameters=self.distortion[1]
-            )
-            d['detector']['distortion'] = dist_d
-        return d
-
-    """
-    ##################### METHODS
-    """
-    def cartToPixel(self, xy_det, pixels=False):
-        """
-        Convert vstacked array or list of [x,y] points in the center-based
-        cartesian frame {Xd, Yd, Zd} to (i, j) edge-based indices
-
-        i is the row index, measured from the upper-left corner
-        j is the col index, measured from the upper-left corner
-
-        if pixels=True, then (i,j) are integer pixel indices.
-        else (i,j) are continuous coords
-        """
-        xy_det = np.atleast_2d(xy_det)
-
-        npts = len(xy_det)
-
-        tmp_ji = xy_det - np.tile(self.corner_ul, (npts, 1))
-        i_pix = -tmp_ji[:, 1] / self.pixel_size_row - 0.5
-        j_pix = tmp_ji[:, 0] / self.pixel_size_col - 0.5
-
-        ij_det = np.vstack([i_pix, j_pix]).T
-        if pixels:
-            ij_det = np.array(np.round(ij_det), dtype=int)
-        return ij_det
-
-    def pixelToCart(self, ij_det):
-        """
-        Convert vstacked array or list of [i,j] pixel indices
-        (or UL corner-based points) and convert to (x,y) in the
-        cartesian frame {Xd, Yd, Zd}
-        """
-        ij_det = np.atleast_2d(ij_det)
-
-        x = (ij_det[:, 1] + 0.5)*self.pixel_size_col\
-            + self.corner_ll[0]
-        y = (self.rows - ij_det[:, 0] - 0.5)*self.pixel_size_row\
-            + self.corner_ll[1]
-        return np.vstack([x, y]).T
-
-    def angularPixelSize(self, xy, rMat_s=None, tVec_s=None, tVec_c=None):
-        """
-        Wraps xrdutil.angularPixelSize
-        """
-        # munge kwargs
-        if rMat_s is None:
-            rMat_s = ct.identity_3x3
-        if tVec_s is None:
-            tVec_s = ct.zeros_3x1
-        if tVec_c is None:
-            tVec_c = ct.zeros_3x1
-
-        # call function
-        ang_ps = xrdutil.angularPixelSize(
-            xy, (self.pixel_size_row, self.pixel_size_col),
-            self.rmat, rMat_s,
-            self.tvec, tVec_s, tVec_c,
-            distortion=self.distortion,
-            beamVec=self.bvec, etaVec=self.evec)
-        return ang_ps
-
-    def clip_to_panel(self, xy, buffer_edges=True):
-        """
-        if self.roi is not None, uses it by default
-
-        TODO: check if need shape kwarg
-        TODO: optimize ROI search better than list comprehension below
-        TODO: panel_buffer can be a 2-d boolean mask, but needs testing
-
-        """
-        xy = np.atleast_2d(xy)
-
-        if self.roi is not None:
-            ij_crds = self.cartToPixel(xy, pixels=True)
-            ii, jj = polygon(self.roi[:, 0], self.roi[:, 1],
-                             shape=(self.rows, self.cols))
-            on_panel_rows = [i in ii for i in ij_crds[:, 0]]
-            on_panel_cols = [j in jj for j in ij_crds[:, 1]]
-            on_panel = np.logical_and(on_panel_rows, on_panel_cols)
-        else:
-            xlim = 0.5*self.col_dim
-            ylim = 0.5*self.row_dim
-            if buffer_edges and self.panel_buffer is not None:
-                if self.panel_buffer.ndim == 2:
-                    pix = self.cartToPixel(xy, pixels=True)
-
-                    roff = np.logical_or(pix[:, 0] < 0, pix[:, 0] >= self.rows)
-                    coff = np.logical_or(pix[:, 1] < 0, pix[:, 1] >= self.cols)
-
-                    idx = np.logical_or(roff, coff)
-
-                    pix[idx, :] = 0
-
-                    on_panel = self.panel_buffer[pix[:, 0], pix[:, 1]]
-                    on_panel[idx] = False
-                else:
-                    xlim -= self.panel_buffer[0]
-                    ylim -= self.panel_buffer[1]
-                    on_panel_x = np.logical_and(
-                        xy[:, 0] >= -xlim, xy[:, 0] <= xlim
-                    )
-                    on_panel_y = np.logical_and(
-                        xy[:, 1] >= -ylim, xy[:, 1] <= ylim
-                    )
-                    on_panel = np.logical_and(on_panel_x, on_panel_y)
-            elif not buffer_edges:
-                    on_panel_x = np.logical_and(
-                        xy[:, 0] >= -xlim, xy[:, 0] <= xlim
-                    )
-                    on_panel_y = np.logical_and(
-                        xy[:, 1] >= -ylim, xy[:, 1] <= ylim
-                    )
-                    on_panel = np.logical_and(on_panel_x, on_panel_y)
-        return xy[on_panel, :], on_panel
-
-    def cart_to_angles(self, xy_data):
-        """
-        TODO: distortion
-        """
-        rmat_s = ct.identity_3x3
-        tvec_s = ct.zeros_3
-        tvec_c = ct.zeros_3
-        angs, g_vec = detectorXYToGvec(
-            xy_data, self.rmat, rmat_s,
-            self.tvec, tvec_s, tvec_c,
-            beamVec=self.bvec, etaVec=self.evec)
-        tth_eta = np.vstack([angs[0], angs[1]]).T
-        return tth_eta, g_vec
-
-    def angles_to_cart(self, tth_eta):
-        """
-        TODO: distortion
-        """
-        rmat_s = rmat_c = ct.identity_3x3
-        tvec_s = tvec_c = ct.zeros_3
-
-        angs = np.hstack([tth_eta, np.zeros((len(tth_eta), 1))])
-
-        xy_det = gvecToDetectorXY(
-            anglesToGVec(angs, bHat_l=self.bvec, eHat_l=self.evec),
-            self.rmat, rmat_s, rmat_c,
-            self.tvec, tvec_s, tvec_c,
-            beamVec=self.bvec)
-        return xy_det
-
-    def interpolate_nearest(self, xy, img, pad_with_nans=True):
-        """
-        TODO: revisit normalization in here?
-
-        """
-        is_2d = img.ndim == 2
-        right_shape = img.shape[0] == self.rows and img.shape[1] == self.cols
-        assert is_2d and right_shape,\
-            "input image must be 2-d with shape (%d, %d)"\
-            % (self.rows, self.cols)
-
-        # initialize output with nans
-        if pad_with_nans:
-            int_xy = np.nan*np.ones(len(xy))
-        else:
-            int_xy = np.zeros(len(xy))
-
-        # clip away points too close to or off the edges of the detector
-        xy_clip, on_panel = self.clip_to_panel(xy, buffer_edges=True)
-
-        # get pixel indices of clipped points
-        i_src = cellIndices(self.row_pixel_vec, xy_clip[:, 1])
-        j_src = cellIndices(self.col_pixel_vec, xy_clip[:, 0])
-
-        # next interpolate across cols
-        int_vals = img[i_src, j_src]
-        int_xy[on_panel] = int_vals
-        return int_xy
-
-    def interpolate_bilinear(self, xy, img, pad_with_nans=True):
-        """
-        TODO: revisit normalization in here?
-        """
-        is_2d = img.ndim == 2
-        right_shape = img.shape[0] == self.rows and img.shape[1] == self.cols
-        assert is_2d and right_shape,\
-            "input image must be 2-d with shape (%d, %d)"\
-            % (self.rows, self.cols)
-
-        # initialize output with nans
-        if pad_with_nans:
-            int_xy = np.nan*np.ones(len(xy))
-        else:
-            int_xy = np.zeros(len(xy))
-
-        # clip away points too close to or off the edges of the detector
-        xy_clip, on_panel = self.clip_to_panel(xy, buffer_edges=True)
-
-        # grab fractional pixel indices of clipped points
-        ij_frac = self.cartToPixel(xy_clip)
-
-        # get floors/ceils from array of pixel _centers_
-        i_floor = cellIndices(self.row_pixel_vec, xy_clip[:, 1])
-        j_floor = cellIndices(self.col_pixel_vec, xy_clip[:, 0])
-        i_ceil = i_floor + 1
-        j_ceil = j_floor + 1
-
-        # first interpolate at top/bottom rows
-        row_floor_int = \
-            (j_ceil - ij_frac[:, 1])*img[i_floor, j_floor] \
-            + (ij_frac[:, 1] - j_floor)*img[i_floor, j_ceil]
-        row_ceil_int = \
-            (j_ceil - ij_frac[:, 1])*img[i_ceil, j_floor] \
-            + (ij_frac[:, 1] - j_floor)*img[i_ceil, j_ceil]
-
-        # next interpolate across cols
-        int_vals = \
-            (i_ceil - ij_frac[:, 0])*row_floor_int \
-            + (ij_frac[:, 0] - i_floor)*row_ceil_int
-        int_xy[on_panel] = int_vals
-        return int_xy
-
-    def make_powder_rings(
-            self, pd, merge_hkls=False, delta_tth=None,
-            delta_eta=10., eta_period=None,
-            rmat_s=ct.identity_3x3,  tvec_s=ct.zeros_3,
-            tvec_c=ct.zeros_3, output_ranges=False, output_etas=False):
-        """
-        """
-        # in case you want to give it tth angles directly
-        if hasattr(pd, '__len__'):
-            tth = np.array(pd).flatten()
-        else:
-            if merge_hkls:
-                tth_idx, tth_ranges = pd.getMergedRanges()
-                if output_ranges:
-                    tth = np.r_[tth_ranges].flatten()
-                else:
-                    tth = np.array([0.5*sum(i) for i in tth_ranges])
-            else:
-                if output_ranges:
-                    tth = pd.getTThRanges().flatten()
-                else:
-                    tth = pd.getTTh()
-
-        if delta_tth is not None:
-            pd.tThWidth = np.radians(delta_tth)
-        else:
-            delta_tth = np.degrees(pd.tThWidth)
-        sector_vec = 0.5*np.radians(
-            [-delta_tth, -delta_eta,
-             -delta_tth,  delta_eta,
-             delta_tth,  delta_eta,
-             delta_tth, -delta_eta]
-        )
-
-        # for generating rings
-        if eta_period is None:
-            eta_period = (-np.pi, np.pi)
-
-        neta = int(360./float(delta_eta))
-        eta = mapAngle(
-            np.radians(delta_eta*(np.linspace(0, neta - 1, num=neta) + 0.5)) +
-            eta_period[0], eta_period
-        )
-
-        angs = [np.vstack([i*np.ones(neta), eta, np.zeros(neta)]) for i in tth]
-
-        # need xy coords and pixel sizes
-        valid_ang = []
-        valid_xy = []
-        map_indices = []
-        for i_ring in range(len(angs)):
-            these_angs = angs[i_ring].T
-            gVec_ring_l = anglesToGVec(these_angs, bHat_l=self.bvec)
-            xydet_ring = gvecToDetectorXY(
-                gVec_ring_l,
-                self.rmat, rmat_s, ct.identity_3x3,
-                self.tvec, tvec_s, tvec_c,
-                beamVec=self.bvec)
-            xydet_ring, on_panel = self.clip_to_panel(xydet_ring)
-            nangs_r = len(xydet_ring)
-
-            # now expand and check to see which sectors (patches) fall off
-            patch_vertices = (
-                np.tile(these_angs[on_panel, :2], (1, 4)) +
-                np.tile(sector_vec, (nangs_r, 1))
-            ).reshape(4*nangs_r, 2)
-
-            # duplicate ome array
-            ome_dupl = np.tile(
-                these_angs[on_panel, 2], (4, 1)
-            ).T.reshape(len(patch_vertices), 1)
-
-            # find vertices that fall on the panel
-            gVec_ring_l = anglesToGVec(
-                np.hstack([patch_vertices, ome_dupl]),
-                bHat_l=self.bvec)
-            tmp_xy = gvecToDetectorXY(
-                gVec_ring_l,
-                self.rmat, rmat_s, ct.identity_3x3,
-                self.tvec, tvec_s, tvec_c,
-                beamVec=self.bvec)
-            tmp_xy, on_panel_p = self.clip_to_panel(tmp_xy)
-
-            # all vertices must be on...
-            patch_is_on = np.all(on_panel_p.reshape(nangs_r, 4), axis=1)
-
-            idx = np.where(on_panel)[0][patch_is_on]
-
-            valid_ang.append(these_angs[idx, :2])
-            valid_xy.append(xydet_ring[patch_is_on])
-            map_indices.append(idx)
-            pass
-        if output_etas:
-            return valid_ang, valid_xy, map_indices, eta
-        else:
-            return valid_ang, valid_xy
-
-    def map_to_plane(self, pts, rmat, tvec):
-        """
-        map detctor points to specified plane
-
-        by convention
-
-        n * (u*pts_l - tvec) = 0
-
-        [pts]_l = rmat*[pts]_m + tvec
-        """
-        # arg munging
-        pts = np.atleast_2d(pts)
-        npts = len(pts)
-
-        # map plane normal & translation vector, LAB FRAME
-        nvec_map_lab = rmat[:, 2].reshape(3, 1)
-        tvec_map_lab = np.atleast_2d(tvec).reshape(3, 1)
-        tvec_d_lab = np.atleast_2d(self.tvec).reshape(3, 1)
-
-        # put pts as 3-d in panel CS and transform to 3-d lab coords
-        pts_det = np.hstack([pts, np.zeros((npts, 1))])
-        pts_lab = np.dot(self.rmat, pts_det.T) + tvec_d_lab
-
-        # scaling along pts vectors to hit map plane
-        u = np.dot(nvec_map_lab.T, tvec_map_lab) \
-            / np.dot(nvec_map_lab.T, pts_lab)
-
-        # pts on map plane, in LAB FRAME
-        pts_map_lab = np.tile(u, (3, 1)) * pts_lab
-
-        return np.dot(rmat.T, pts_map_lab - tvec_map_lab)[:2, :].T
-
-    def simulate_rotation_series(self, plane_data, grain_param_list,
-                                 eta_ranges=[(-np.pi, np.pi), ],
-                                 ome_ranges=[(-np.pi, np.pi), ],
-                                 ome_period=(-np.pi, np.pi),
-                                 chi=0., tVec_s=ct.zeros_3,
-                                 wavelength=None):
-        """
-        """
-
-        # grab B-matrix from plane data
-        bMat = plane_data.latVecOps['B']
-
-        # reconcile wavelength
-        #   * added sanity check on exclusions here; possible to
-        #   * make some reflections invalid (NaN)
-        if wavelength is None:
-            wavelength = plane_data.wavelength
-        else:
-            if plane_data.wavelength != wavelength:
-                plane_data.wavelength = ct.keVToAngstrom(wavelength)
-        assert not np.any(np.isnan(plane_data.getTTh())),\
-            "plane data exclusions incompatible with wavelength"
-
-        # vstacked G-vector id, h, k, l
-        full_hkls = xrdutil._fetch_hkls_from_planedata(plane_data)
-
-        """ LOOP OVER GRAINS """
-        valid_ids = []
-        valid_hkls = []
-        valid_angs = []
-        valid_xys = []
-        ang_pixel_size = []
-        for gparm in grain_param_list:
-
-            # make useful parameters
-            rMat_c = makeRotMatOfExpMap(gparm[:3])
-            tVec_c = gparm[3:6]
-            vInv_s = gparm[6:]
-
-            # All possible bragg conditions as vstacked [tth, eta, ome]
-            # for each omega solution
-            angList = np.vstack(
-                oscillAnglesOfHKLs(
-                    full_hkls[:, 1:], chi,
-                    rMat_c, bMat, wavelength,
-                    vInv=vInv_s,
-                    )
-                )
-
-            # filter by eta and omega ranges
-            # ??? get eta range from detector?
-            allAngs, allHKLs = xrdutil._filter_hkls_eta_ome(
-                full_hkls, angList, eta_ranges, ome_ranges
-                )
-            allAngs[:, 2] = mapAngle(allAngs[:, 2], ome_period)
-
-            # find points that fall on the panel
-            det_xy, rMat_s = xrdutil._project_on_detector_plane(
-                allAngs,
-                self.rmat, rMat_c, chi,
-                self.tvec, tVec_c, tVec_s,
-                self.distortion)
-            xys_p, on_panel = self.clip_to_panel(det_xy)
-            valid_xys.append(xys_p)
-
-            # grab hkls and gvec ids for this panel
-            valid_hkls.append(allHKLs[on_panel, 1:])
-            valid_ids.append(allHKLs[on_panel, 0])
-
-            # reflection angles (voxel centers) and pixel size in (tth, eta)
-            valid_angs.append(allAngs[on_panel, :])
-            ang_pixel_size.append(self.angularPixelSize(xys_p))
-        return valid_ids, valid_hkls, valid_angs, valid_xys, ang_pixel_size
-
-    def simulate_laue_pattern(self, crystal_data,
-                              minEnergy=5., maxEnergy=35.,
-                              rmat_s=None, tvec_s=None,
-                              grain_params=None,
-                              beam_vec=None):
-        """
-        """
-        if isinstance(crystal_data, PlaneData):
-
-            plane_data = crystal_data
-
-            # grab the expanded list of hkls from plane_data
-            hkls = np.hstack(plane_data.getSymHKLs())
-
-            # and the unit plane normals (G-vectors) in CRYSTAL FRAME
-            gvec_c = np.dot(plane_data.latVecOps['B'], hkls)
-        elif len(crystal_data) == 2:
-            # !!! should clean this up
-            hkls = np.array(crystal_data[0])
-            bmat = crystal_data[1]
-            gvec_c = np.dot(bmat, hkls)
-        else:
-            raise(RuntimeError, 'argument list not understood')
-        nhkls_tot = hkls.shape[1]
-
-        # parse energy ranges
-        # TODO: allow for spectrum parsing
-        multipleEnergyRanges = False
-        if hasattr(maxEnergy, '__len__'):
-            assert len(maxEnergy) == len(minEnergy), \
-                'energy cutoff ranges must have the same length'
-            multipleEnergyRanges = True
-            lmin = []
-            lmax = []
-            for i in range(len(maxEnergy)):
-                lmin.append(ct.keVToAngstrom(maxEnergy[i]))
-                lmax.append(ct.keVToAngstrom(minEnergy[i]))
-        else:
-            lmin = ct.keVToAngstrom(maxEnergy)
-            lmax = ct.keVToAngstrom(minEnergy)
-
-        # parse grain parameters kwarg
-        if grain_params is None:
-            grain_params = np.atleast_2d(
-                np.hstack([np.zeros(6), ct.identity_6x1])
-            )
-        n_grains = len(grain_params)
-
-        # sample rotation
-        if rmat_s is None:
-            rmat_s = ct.identity_3x3
-
-        # dummy translation vector... make input
-        if tvec_s is None:
-            tvec_s = ct.zeros_3
-
-        # beam vector
-        if beam_vec is None:
-            beam_vec = ct.beam_vec
-
-        # =========================================================================
-        # LOOP OVER GRAINS
-        # =========================================================================
-
-        # pre-allocate output arrays
-        xy_det = np.nan*np.ones((n_grains, nhkls_tot, 2))
-        hkls_in = np.nan*np.ones((n_grains, 3, nhkls_tot))
-        angles = np.nan*np.ones((n_grains, nhkls_tot, 2))
-        dspacing = np.nan*np.ones((n_grains, nhkls_tot))
-        energy = np.nan*np.ones((n_grains, nhkls_tot))
-        for iG, gp in enumerate(grain_params):
-            rmat_c = makeRotMatOfExpMap(gp[:3])
-            tvec_c = gp[3:6].reshape(3, 1)
-            vInv_s = mutil.vecMVToSymm(gp[6:].reshape(6, 1))
-
-            # stretch them: V^(-1) * R * Gc
-            gvec_s_str = np.dot(vInv_s, np.dot(rmat_c, gvec_c))
-            ghat_c_str = mutil.unitVector(np.dot(rmat_c.T, gvec_s_str))
-
-            # project
-            dpts = gvecToDetectorXY(ghat_c_str.T,
-                                    self.rmat, rmat_s, rmat_c,
-                                    self.tvec, tvec_s, tvec_c,
-                                    beamVec=beam_vec)
-
-            # check intersections with detector plane
-            canIntersect = ~np.isnan(dpts[:, 0])
-            npts_in = sum(canIntersect)
-
-            if np.any(canIntersect):
-                dpts = dpts[canIntersect, :].reshape(npts_in, 2)
-                dhkl = hkls[:, canIntersect].reshape(3, npts_in)
-
-                # back to angles
-                tth_eta, gvec_l = detectorXYToGvec(
-                    dpts,
-                    self.rmat, rmat_s,
-                    self.tvec, tvec_s, tvec_c,
-                    beamVec=beam_vec)
-                tth_eta = np.vstack(tth_eta).T
-
-                # warp measured points
-                if self.distortion is not None:
-                    if len(self.distortion) == 2:
-                        dpts = self.distortion[0](
-                            dpts, self.distortion[1],
-                            invert=True)
-                    else:
-                        raise(RuntimeError,
-                              "something is wrong with the distortion")
-
-                # plane spacings and energies
-                dsp = 1. / rowNorm(gvec_s_str[:, canIntersect].T)
-                wlen = 2*dsp*np.sin(0.5*tth_eta[:, 0])
-
-                # clip to detector panel
-                _, on_panel = self.clip_to_panel(dpts, buffer_edges=True)
-
-                if multipleEnergyRanges:
-                    validEnergy = np.zeros(len(wlen), dtype=bool)
-                    for i in range(len(lmin)):
-                        in_energy_range = np.logical_and(
-                                wlen >= lmin[i],
-                                wlen <= lmax[i])
-                        validEnergy = validEnergy | in_energy_range
-                        pass
-                else:
-                    validEnergy = np.logical_and(wlen >= lmin, wlen <= lmax)
-                    pass
-
-                # index for valid reflections
-                keepers = np.where(np.logical_and(on_panel, validEnergy))[0]
-
-                # assign output arrays
-                xy_det[iG][keepers, :] = dpts[keepers, :]
-                hkls_in[iG][:, keepers] = dhkl[:, keepers]
-                angles[iG][keepers, :] = tth_eta[keepers, :]
-                dspacing[iG, keepers] = dsp[keepers]
-                energy[iG, keepers] = ct.keVToAngstrom(wlen[keepers])
-                pass    # close conditional on valids
-            pass    # close loop on grains
-        return xy_det, hkls_in, angles, dspacing, energy
-
-# =============================================================================
-# UTILITIES
-# =============================================================================
-
-
-class PatchDataWriter(object):
-    """
-    """
-    def __init__(self, filename):
-        self._delim = '  '
-        header_items = (
-            '# ID', 'PID',
-            'H', 'K', 'L',
-            'sum(int)', 'max(int)',
-            'pred tth', 'pred eta', 'pred ome',
-            'meas tth', 'meas eta', 'meas ome',
-            'pred X', 'pred Y',
-            'meas X', 'meas Y'
-        )
-        self._header = self._delim.join([
-            self._delim.join(np.tile('{:<6}', 5)).format(*header_items[:5]),
-            self._delim.join(np.tile('{:<12}', 2)).format(*header_items[5:7]),
-            self._delim.join(np.tile('{:<23}', 10)).format(*header_items[7:17])
-        ])
-        if isinstance(filename, file):
-            self.fid = filename
-        else:
-            self.fid = open(filename, 'w')
-        print(self._header, file=self.fid)
-
-    def __del__(self):
-        self.close()
-
-    def close(self):
-        self.fid.close()
-
-    def dump_patch(self, peak_id, hkl_id,
-                   hkl, spot_int, max_int,
-                   pangs, mangs, pxy, mxy):
-        """
-        !!! maybe need to check that last four inputs are arrays
-        """
-        if mangs is None:
-            spot_int = np.nan
-            max_int = np.nan
-            mangs = np.ones(3)*np.nan
-            mxy = np.ones(2)*np.nan
-
-        res = [int(peak_id), int(hkl_id)] \
-            + np.array(hkl, dtype=int).tolist() \
-            + [spot_int, max_int] \
-            + pangs.tolist() \
-            + mangs.tolist() \
-            + pxy.tolist() \
-            + mxy.tolist()
-
-        output_str = self._delim.join(
-            [self._delim.join(np.tile('{:<6d}', 5)).format(*res[:5]),
-             self._delim.join(np.tile('{:<12e}', 2)).format(*res[5:7]),
-             self._delim.join(np.tile('{:<23.16e}', 10)).format(*res[7:])]
-        )
-        print(output_str, file=self.fid)
-        return output_str
-
-
-class GrainDataWriter(object):
-    """
-    """
-    def __init__(self, filename):
-        self._delim = '  '
-        header_items = (
-            '# grain ID', 'completeness', 'chi^2',
-            'exp_map_c[0]', 'exp_map_c[1]', 'exp_map_c[2]',
-            't_vec_c[0]', 't_vec_c[1]', 't_vec_c[2]',
-            'inv(V_s)[0,0]', 'inv(V_s)[1,1]', 'inv(V_s)[2,2]',
-            'inv(V_s)[1,2]*sqrt(2)',
-            'inv(V_s)[0,2]*sqrt(2)',
-            'inv(V_s)[0,2]*sqrt(2)',
-            'ln(V_s)[0,0]', 'ln(V_s)[1,1]', 'ln(V_s)[2,2]',
-            'ln(V_s)[1,2]', 'ln(V_s)[0,2]', 'ln(V_s)[0,1]'
-        )
-        self._header = self._delim.join(
-            [self._delim.join(
-                np.tile('{:<12}', 3)
-                ).format(*header_items[:3]),
-             self._delim.join(
-                np.tile('{:<23}', len(header_items) - 3)
-                ).format(*header_items[3:])]
-        )
-        if isinstance(filename, file):
-            self.fid = filename
-        else:
-            self.fid = open(filename, 'w')
-        print(self._header, file=self.fid)
-
-    def __del__(self):
-        self.close()
-
-    def close(self):
-        self.fid.close()
-
-    def dump_grain(self, grain_id, completeness, chisq,
-                   grain_params):
-        assert len(grain_params) == 12, \
-            "len(grain_params) must be 12, not %d" % len(grain_params)
-
-        # extract strain
-        emat = logm(np.linalg.inv(mutil.vecMVToSymm(grain_params[6:])))
-        evec = mutil.symmToVecMV(emat, scale=False)
-
-        res = [int(grain_id), completeness, chisq] \
-            + grain_params.tolist() \
-            + evec.tolist()
-        output_str = self._delim.join(
-            [self._delim.join(
-                ['{:<12d}', '{:<12f}', '{:<12e}']
-             ).format(*res[:3]),
-             self._delim.join(
-                np.tile('{:<23.16e}', len(res) - 3)
-             ).format(*res[3:])]
-        )
-        print(output_str, file=self.fid)
-        return output_str
-
-
-class GrainDataWriter_h5(object):
-    """
-    TODO: add material spec
-    """
-    def __init__(self, filename, instr_cfg, grain_params, use_attr=False):
-
-        if isinstance(filename, h5py.File):
-            self.fid = filename
-        else:
-            self.fid = h5py.File(filename + ".hdf5", "w")
-        icfg = {}
-        icfg.update(instr_cfg)
-
-        # add instrument groups and attributes
-        self.instr_grp = self.fid.create_group('instrument')
-        unwrap_dict_to_h5(self.instr_grp, icfg, asattr=use_attr)
-
-        # add grain group
-        self.grain_grp = self.fid.create_group('grain')
-        rmat_c = makeRotMatOfExpMap(grain_params[:3])
-        tvec_c = np.array(grain_params[3:6]).flatten()
-        vinv_s = np.array(grain_params[6:]).flatten()
-        vmat_s = np.linalg.inv(mutil.vecMVToSymm(vinv_s))
-
-        if use_attr:    # attribute version
-            self.grain_grp.attrs.create('rmat_c', rmat_c)
-            self.grain_grp.attrs.create('tvec_c', tvec_c)
-            self.grain_grp.attrs.create('inv(V)_s', vinv_s)
-            self.grain_grp.attrs.create('vmat_s', vmat_s)
-        else:    # dataset version
-            self.grain_grp.create_dataset('rmat_c', data=rmat_c)
-            self.grain_grp.create_dataset('tvec_c', data=tvec_c)
-            self.grain_grp.create_dataset('inv(V)_s', data=vinv_s)
-            self.grain_grp.create_dataset('vmat_s', data=vmat_s)
-
-        data_key = 'reflection_data'
-        self.data_grp = self.fid.create_group(data_key)
-
-        for det_key in self.instr_grp['detectors'].keys():
-            self.data_grp.create_group(det_key)
-
-    # FIXME: throws exception when called after close method
-    # def __del__(self):
-    #    self.close()
-
-    def close(self):
-        self.fid.close()
-
-    def dump_patch(self, panel_id,
-                   i_refl, peak_id, hkl_id, hkl,
-                   tth_edges, eta_edges, ome_centers,
-                   xy_centers, ijs, frame_indices,
-                   spot_data, pangs, pxy, mangs, mxy, gzip=1):
-        """
-        to be called inside loop over patches
-
-        default GZIP level for data arrays is 1
-        """
-        fi = np.array(frame_indices, dtype=int)
-
-        panel_grp = self.data_grp[panel_id]
-        spot_grp = panel_grp.create_group("spot_%05d" % i_refl)
-        spot_grp.attrs.create('peak_id', int(peak_id))
-        spot_grp.attrs.create('hkl_id', int(hkl_id))
-        spot_grp.attrs.create('hkl', np.array(hkl, dtype=int))
-        spot_grp.attrs.create('predicted_angles', pangs)
-        spot_grp.attrs.create('predicted_xy', pxy)
-        if mangs is None:
-            mangs = np.nan*np.ones(3)
-        spot_grp.attrs.create('measured_angles', mangs)
-        if mxy is None:
-            mxy = np.nan*np.ones(3)
-        spot_grp.attrs.create('measured_xy', mxy)
-
-        # get centers crds from edge arrays
-        # FIXME: export full coordinate arrays, or just center vectors???
-        #
-        # ome_crd, eta_crd, tth_crd = np.meshgrid(
-        #     ome_centers,
-        #     centers_of_edge_vec(eta_edges),
-        #     centers_of_edge_vec(tth_edges),
-        #     indexing='ij')
-        #
-        # ome_dim, eta_dim, tth_dim = spot_data.shape
-
-        # !!! for now just exporting center vectors for spot_data
-        tth_crd = centers_of_edge_vec(tth_edges)
-        eta_crd = centers_of_edge_vec(eta_edges)
-
-        shuffle_data = True  # reduces size by 20%
-        spot_grp.create_dataset('tth_crd', data=tth_crd,
-                                compression="gzip", compression_opts=gzip,
-                                shuffle=shuffle_data)
-        spot_grp.create_dataset('eta_crd', data=eta_crd,
-                                compression="gzip", compression_opts=gzip,
-                                shuffle=shuffle_data)
-        spot_grp.create_dataset('ome_crd', data=ome_centers,
-                                compression="gzip", compression_opts=gzip,
-                                shuffle=shuffle_data)
-        spot_grp.create_dataset('xy_centers', data=xy_centers,
-                                compression="gzip", compression_opts=gzip,
-                                shuffle=shuffle_data)
-        spot_grp.create_dataset('ij_centers', data=ijs,
-                                compression="gzip", compression_opts=gzip,
-                                shuffle=shuffle_data)
-        spot_grp.create_dataset('frame_indices', data=fi,
-                                compression="gzip", compression_opts=gzip,
-                                shuffle=shuffle_data)
-        spot_grp.create_dataset('intensities', data=spot_data,
-                                compression="gzip", compression_opts=gzip,
-                                shuffle=shuffle_data)
-        return
-
-
-def unwrap_dict_to_h5(grp, d, asattr=True):
-    while len(d) > 0:
-        key, item = d.popitem()
-        if isinstance(item, dict):
-            subgrp = grp.create_group(key)
-            unwrap_dict_to_h5(subgrp, item)
-        else:
-            if asattr:
-                grp.attrs.create(key, item)
-            else:
-                grp.create_dataset(key, data=np.atleast_1d(item))
-
-
-class GenerateEtaOmeMaps(object):
-    """
-    eta-ome map class derived from new image_series and YAML config
-
-    ...for now...
-
-    must provide:
-
-    self.dataStore
-    self.planeData
-    self.iHKLList
-    self.etaEdges # IN RADIANS
-    self.omeEdges # IN RADIANS
-    self.etas     # IN RADIANS
-    self.omegas   # IN RADIANS
-
-    """
-    def __init__(self, image_series_dict, instrument, plane_data,
-                 active_hkls=None, eta_step=0.25, threshold=None,
-                 ome_period=(0, 360)):
-        """
-        image_series must be OmegaImageSeries class
-        instrument_params must be a dict (loaded from yaml spec)
-        active_hkls must be a list (required for now)
-        """
-
-        self._planeData = plane_data
-
-        # ???: change name of iHKLList?
-        # ???: can we change the behavior of iHKLList?
-        if active_hkls is None:
-            n_rings = len(plane_data.getTTh())
-            self._iHKLList = range(n_rings)
-        else:
-            self._iHKLList = active_hkls
-            n_rings = len(active_hkls)
-
-        # ???: need to pass a threshold?
-        eta_mapping, etas = instrument.extract_polar_maps(
-            plane_data, image_series_dict,
-            active_hkls=active_hkls, threshold=threshold,
-            tth_tol=None, eta_tol=eta_step)
-
-        # grab a det key
-        # WARNING: this process assumes that the imageseries for all panels
-        # have the same length and omegas
-        det_key = eta_mapping.keys()[0]
-        data_store = []
-        for i_ring in range(n_rings):
-            full_map = np.zeros_like(eta_mapping[det_key][i_ring])
-            nan_mask_full = np.zeros(
-                (len(eta_mapping), full_map.shape[0], full_map.shape[1])
-            )
-            i_p = 0
-            for det_key, eta_map in eta_mapping.iteritems():
-                nan_mask = ~np.isnan(eta_map[i_ring])
-                nan_mask_full[i_p] = nan_mask
-                full_map[nan_mask] += eta_map[i_ring][nan_mask]
-                i_p += 1
-            re_nan_these = np.sum(nan_mask_full, axis=0) == 0
-            full_map[re_nan_these] = np.nan
-            data_store.append(full_map)
-        self._dataStore = data_store
-
-        # handle omegas
-        omegas_array = image_series_dict[det_key].metadata['omega']
-        self._omegas = mapAngle(
-            np.radians(np.average(omegas_array, axis=1)),
-            np.radians(ome_period)
-        )
-        self._omeEdges = mapAngle(
-            np.radians(np.r_[omegas_array[:, 0], omegas_array[-1, 1]]),
-            np.radians(ome_period)
-        )
-
-        # handle etas
-        # WARNING: unlinke the omegas in imageseries metadata,
-        # these are in RADIANS and represent bin centers
-        self._etas = etas
-        self._etaEdges = np.r_[
-            etas - 0.5*np.radians(eta_step),
-            etas[-1] + 0.5*np.radians(eta_step)]
-
-    @property
-    def dataStore(self):
-        return self._dataStore
-
-    @property
-    def planeData(self):
-        return self._planeData
-
-    @property
-    def iHKLList(self):
-        return np.atleast_1d(self._iHKLList).flatten()
-
-    @property
-    def etaEdges(self):
-        return self._etaEdges
-
-    @property
-    def omeEdges(self):
-        return self._omeEdges
-
-    @property
-    def etas(self):
-        return self._etas
-
-    @property
-    def omegas(self):
-        return self._omegas
-
-    def save(self, filename):
-        """
-        self.dataStore
-        self.planeData
-        self.iHKLList
-        self.etaEdges
-        self.omeEdges
-        self.etas
-        self.omegas
-        """
-        args = np.array(self.planeData.getParams())[:4]
-        args[2] = valWUnit('wavelength', 'length', args[2], 'angstrom')
-        hkls = self.planeData.hkls
-        save_dict = {'dataStore': self.dataStore,
-                     'etas': self.etas,
-                     'etaEdges': self.etaEdges,
-                     'iHKLList': self.iHKLList,
-                     'omegas': self.omegas,
-                     'omeEdges': self.omeEdges,
-                     'planeData_args': args,
-                     'planeData_hkls': hkls}
-        np.savez_compressed(filename, **save_dict)
-        return
-    pass  # end of class: GenerateEtaOmeMaps
